feat: Add component to Header
