+++ src/views/common/Header.vue
@@ -1,0 +1,40 @@
+<template>
+  <div class="header">
+    <h1 class="header__logo">Style guide views</h1>
+    <GlobalNavi v-if="loggedIn" />
+  </div>
+</template>
+
+<script lang="ts">
+import { Component, Prop, Vue } from "vue-property-decorator";
+import GlobalNavi from "@/views/common/GlobalNavi.vue";
+
+@Component({
+  components: {
+    GlobalNavi
+  }
+})
+export default class Header extends Vue {
+  @Prop({ type: Boolean, required: true })
+  loggedIn!: boolean;
+}
+</script>
+
+<style scoped lang="scss">
+.header {
+  background-color: $is_color_white;
+}
+
+.header__logo {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: $header_height;
+  background: linear-gradient(180deg, $is_base_color300, $is_base_color100);
+  background-clip: text; //テキストでくり抜く
+  color: $is_base_color200;
+
+  -webkit-text-fill-color: transparent;
+  @include font-size(30);
+}
+</style>
